From 04825384999f81c08e820fe380e95b6325352a6d Mon Sep 17 00:00:00 2001 From: Rohan Garg <rohan.garg@collabora.com> Date: Mon, 12 Aug 2019 17:04:34 +0200 Subject: [PATCH] rockchip: rk3399: derive ethaddr from cpuid
Generate a MAC address based on the cpuid available in the efuse
block: Use the first 6 byte of the cpuid's SHA256 hash and set the
locally administered bits. Also ensure that the multicast bit is
cleared.

The MAC address is only generated and set if there is no ethaddr
present in the saved environment.

This is based off of Klaus Goger's work in 8adc9d

Signed-off-by: Rohan Garg <rohan.garg@collabora.com>
Reviewed-by: Kever Yang <kever.yang@rock-chips.com>
(cherry picked from commit 10d74ac64a662ba2771bb297470508517a9e3af5)
Signed-off-by: Stephen <stephen@vamrs.com>