configs: khadas-viml3_android_ab: set CONFIG_SYS_MMC_ENV_{DEV,PART}
In commit 7d080773347c ("Convert CONFIG_SYS_MMC_ENV_DEV et al to
Kconfig"), the mmc env hardware number and partition number were moved
to a KConfig option.

Since khadas-vim3_android_ab is out of tree, this was not applied on
the config.

Add the same options as the ones present in
configs/khadas-vim3l_defconfig

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>